+++ src/ts/Ads/Views/OpenMeasurement/OpenMeasurement.ts
@@ -439,8 +439,11 @@ export class OpenMeasurement<T extends Campaign> extends View<T> {
                 SDKMetrics.reportMetricEvent(OMMetric.IASVerificatonInjected);
             }
 
-            if (vendorKey.startsWith('doubleclickbygoogle.com') && this._campaign instanceof AdMobCampaign) {
-                SDKMetrics.reportMetricEvent(AdmobMetric.DoubleClickOMInjections);
+            if (OpenMeasurementUtilities.isDoubleClickGoogle(vendorKey) && this._campaign instanceof AdMobCampaign) {
+                SDKMetrics.reportMetricEventWithTags(AdmobMetric.DoubleClickOMInjections,
+                [
+                    SDKMetrics.createAdsSdkTag('dckey', OpenMeasurementUtilities.getDcKeyMetricTag(vendorKey))
+                ]);
             }
 
             return Promise.resolve();